% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir2.R
\name{dir2}
\alias{dir2}
\title{List files and directories recursively to a certain depth}
\usage{
dir2(path = ".", pattern = NULL, all.files = FALSE, full.names = TRUE,
  recursive = FALSE, ignore.case = FALSE, include.dirs = FALSE,
  absolute = FALSE)
}
\arguments{
\item{path}{a character vector of paths.}

\item{pattern}{(optional) an optional regular expression.
Full path names (including path and file names) that match the
regular expression will be returned.}

\item{all.files}{If \code{FALSE}, non-visible files (prefixed with a \code{.}) are
not returned.  If \code{TRUE}, all files are returned.}

\item{full.names}{If \code{TRUE} or \code{recursive = TRUE} (or a positive number),
full path names are returned, otherwise only the file names.}

\item{recursive}{If \code{FALSE} or \code{0}, content of sub-directories is not
included. if \code{TRUE} or \code{+Inf}, content of all sub-directories are
recursively included.  If a positive number \code{depth}, then
sub-directories to that depth is recursively included.}

\item{ignore.case}{If \code{TRUE}, pattern matching is case-insensitive,
othewise not.}

\item{include.dirs}{If \code{TRUE}, also directories are returned, otherwise not.}

\item{absolute}{If \code{TRUE} or \code{length(path) > 1}, absolute path names are
returned and used in the file pattern matching.}
}
\value{
Character vector of identified files and directories.
}
\description{
List files and directories recursively to a certain depth
}
\details{
Files \file{.} and \file{..} (the current and parent directories) are never
returned.
}
\examples{
path <- system.file(package = "dirdf")
files <- dir2(path, include.dirs = TRUE)
print(files)

files <- dir2(path, recursive = 1L)
print(files)

files <- dir2(path, recursive = TRUE)
print(files)
}
\seealso{
\code{\link[base:dir]{base::dir()}}
}
