% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirdf.R
\name{dirdf}
\alias{dirdf}
\title{Creates a data frame using information from paths and file names.
It searches through the directories in order to create the path names of the files.}
\usage{
dirdf(paths, template = NULL, regexp = NULL, colnames = NULL,
  missing = NA_character_, recursive = TRUE, ...)
}
\arguments{
\item{paths}{character vector with zero or more paths that will be searched}

\item{template}{character string containing the path that will be used to create the column names and the regular expression to parse the file names.}

\item{regexp}{regular expression used to parse the file names}

\item{colnames}{vector containing the names of the columns in the data frame}

\item{missing}{value for a missing variable value}

\item{recursive}{if TRUE, it will recursively search over directories}

\item{...}{Additional arguments pass to \code{\link[base]{dir}()}.}
}
\description{
It accepts either a template or a regular expression and column names.
}
\examples{
\dontrun{
example_path <- system.file(package = "dirdf", "examples", "dataset_1")
example_template1 <- "Year-Month-Day_Assay_Plasmid-Type-Fraction_WellNumber.extension"
example_template2 <- "Date_Assay_Experiment_WellNumber.extension"

dirdf(example_path1, template = example_template)
}

}

