% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirdf_parse.R
\name{dirdf_parse}
\alias{dirdf_parse}
\title{Creates a data frame using information from the paths and file names.
It accepts either a template or a regular expression and column names.}
\usage{
dirdf_parse(pathnames, template = NULL, regexp = NULL, colnames = NULL,
  missing = NA_character_)
}
\arguments{
\item{pathnames}{character vector 0 or more pathnames}

\item{template}{template character string containing the path that will be used to create the column names and the regular expression to parse the file names.}

\item{regexp}{regular expression used to parse the file names}

\item{colnames}{colnames vector containing the names of the columns in the data frame}

\item{missing}{missing value for a missing variable value}
}
\description{
Creates a data frame using information from the paths and file names.
It accepts either a template or a regular expression and column names.
}
\examples{
\dontrun{
example_pathnames <- dir(system.file(package = "dirdf", "examples", "dataset_1"))

example_template1 <- "Year-Month-Day_Assay_Plasmid-Type-Fraction_WellNumber.extension"
example_template2 <- "Date_Assay_Experiment_WellNumber.extension"
}

}

